wibl-python/wibl-manager/requirements.txt
gunicorn~=23.0.0
Flask~=2.2.3
Flask-SQLAlchemy~=3.0.3
Flask-RESTful~=0.3.9
requests~=2.32.4
dataclasses-json~=0.5.7
boto3~=1.34.0
psycopg2-binary~=2.9.9
alembic
